+++ satella/coding/sequences/iterators.py
@@ -1,5 +1,6 @@
 import logging
 import typing as tp
+import warnings
 
 logger = logging.getLogger(__name__)
 
@@ -11,7 +12,7 @@ def infinite_counter(start_at: int = 0, step: int = 1) -> tp.Generator[int, None
     :param start_at: value at which to start counting. It will be yielded as first
     :param step: step by which to progress the counter
     """
-
+    warnings.warn('This is deprecated, use itertools.count() instead', DeprecationWarning)
     i = start_at
     while True:
         yield i