+++ satella/coding/sequences/__init__.py
@@ -1,4 +1,4 @@
 from .choose import choose
-from .iterators import infinite_counter, take_n
+from .iterators import infinite_counter, take_n, is_instance
 
-__all__ = ['choose', 'infinite_counter', 'take_n']
+__all__ = ['choose', 'infinite_counter', 'take_n', 'is_instance']