+++ tests/test_coding/test_sequences.py
@@ -1,7 +1,7 @@
 import logging
 import unittest
 
-from satella.coding.sequences import choose, infinite_counter, take_n
+from satella.coding.sequences import choose, infinite_counter, take_n, is_instance
 
 logger = logging.getLogger(__name__)
 
@@ -23,3 +23,7 @@ class TestSequences(unittest.TestCase):
         self.assertEqual(choose(lambda x: x == 2, [1, 2, 3, 4, 5]), 2)
         self.assertRaises(ValueError, lambda: choose(lambda x: x % 2 == 0, [1, 2, 3, 4, 5]))
         self.assertRaises(ValueError, lambda: choose(lambda x: x == 0, [1, 2, 3, 4, 5]))
+
+    def test_is_instance(self):
+        objects = [object(), object(), [], [], object()]
+        self.assertEqual(len(list(filter(is_instance(list), objects))), 2)